+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- BlackboardPropertyName now uses FNV-1a hash algorithm.
+
 ## [1.4.0] - 2021-03-12
 
 ### Added